Set BASE to "http://psycoti.ca/0/"
I'm not sure if this was the right thing to do.
I expect it's probably going to make local debugging a little tough.

Basically, the problem I was having was that I was using the same templates in a bunch of directories.
Stuff was ending up all over the place, and the relative links always made me feel a little weird.

I'm not sure if this is better, though...