Added --follow to default.augmented.do
This means it follows file renames (as best it can).
I figured I shouldn't need that in general, since the filenames
are IDs and shouldn't change.

In this case, though, for the port I obviously had to rename files.
And, in general, if a file does change ID for some reason, it's fine
that it tries to preserve the original published date.