Factored Out Header Manipulation into extract_headers
I had a bug in the way I was extracting headers.
Notably, when a file contained lines that looked like headers in the
body of the file, it would extract those too.

I was extracting headers all over the place, and it made me sad.

So, I wrote ./extract_headers to do that for me, and now everything
just uses that.

Happier.