Added text/prose Type
The <pre> </pre> bits weren't really what I wanted.
This is much better.

I turn each group separated by a blank newline as a stanza and then each
source line in that becomes a line in the output.

It preserves the bits of the formatting I care about, while allowing the
browser to still do things with it.

It also ends up with a lot more metadata about stanzas and lines...
That's good, right?

I converted "Someone I Want" over to this new type, and added the css
required to make it formatted like I want.